Extract zoom logic into hooks
